- renaming s3 package to aws helpers in preperation for ssm client
